pogo_transport: Init pogo irqs at the end of probe
There is a possibility that the POGO OVP will be disabled if the device
is docked and then boots up. This is because the status irq is requested
before the initialization of the POGO OVP (which is disabled later in
the probe).

pogo_transport probe    status irq handler        pogo event
--------------------    ------------------        ------------
irq requested
 |--> irq triggerd -->  ovp_en_gpio not init'd;
 |                      skip enabling of the OVP
 |                       |--> queue docking event
ovp_en_gpio initialized        | (delayed work)
 |                             |
ovp_en_gpio pulled high        |
(disabled)                     |
                               |--> timeout  -->  OVP still disabled
			                          voltage 0; retry
						  voltage 0; retry
						  ...

Solve this by moving the irq request to the end of the probe to make
sure that OVP is initialized and ready before the irq is requested.

Bug: 246445724
Signed-off-by: Kyle Tso <kyletso@google.com>
Change-Id: Ic93a21ccddcb6aa93f6786228a446382f4e46ace