pogo_transport: Update the CC orientation to SS PHY in some cases
The reg base of the Superspeed PHY is based on the CC orientation. If
the host mode is enabled before CC orientation is decided, the
orientation information should be updated to the Superspeed PHY.
Register a callback in Type-C SW stack to receive the CC-change event.
When the condition is satified, update the CC orientation through extcon
and cast a vote to "SSPHY_RESTART" votable to trigger the PHY restart.

Bug: 257490462
Change-Id: I955ff6705482a4ce12b8233f0e36d185ad7022cd
Signed-off-by: Kyle Tso <kyletso@google.com>