pogo_transport: Update the orientation before enabling USB Data
There is a possibility that when doing the phy restart due to
orientation change, the USB phy is being turned on so that the memory
access causes Asychronous SError Interrupt. Fix this by moving the
orientation updating to a dedicated function and call it prior to the
enabling of the USB phy. Also remove the phy restart during
switch_to_usbc_locked as the updated orientation will be adopted in the
phy in enable_data_path_locked automatically.

Bug: 271689070
Change-Id: I68d95a3f0b0e200bb6503c073ebe708a2102502e
Signed-off-by: Kyle Tso <kyletso@google.com>