rt4539: i2c: Make i2c remove callback return void
The upstream commit ed5c2f5fd10d ("i2c: Make remove callback return
void") changes the return value for the i2c_device remove callback to be
void. So we need to update all of these drivers accordingly. The reason
was because the return value wasn't really being used which is
consistent with this driver as well.

Bug: 289529571
Change-Id: Ib88f0b13d3e044b940f9041fd03d828e6b563436
Signed-off-by: Samuel Gosselin <sgosselin@google.com>