tangorpro: Swap synaptics with broadcom nitrous.ko
These are essentially the same underlying chips, and their modules are
nearly identical, modulo a single (device-tree configured) feature. We'd
like to avoid maintaining two conflicting modules. Since everyone else
is using the broadcom variant, let's follow suit.

To maintain feature parity: drop "cloudripper-bluetooth" from the tangorpro /
synaptics variant; this is simply a modification of the common bluetooth
configuration that adds the "timesync-gpios" property and associated pin
control. This features was not present in the synaptics driver variant,
so probably shouldn't be in the device tree.

When building the dtbos before and after this change, the resulting diff
is:
(a) mostly changes in binary details (renumbering of fragment@XX nodes,
    renumbering of phandles, ,...); and otherwise
(b) the following diff:

 			btbcm {
 				compatible = "goog,nitrous";
-				pinctrl-0 = <0x05 0x06 0x07 0x08>;
+				pinctrl-0 = <0x05 0x06 0x07>;
 				pinctrl-names = "default";
 				shutdown-gpios = <0xffffffff 0x02 0x00>;
 				host-wakeup-gpios = <0xffffffff 0x06 0x00>;
 				device-wakeup-gpios = <0xffffffff 0x03 0x00>;
 				status = "okay";
-				timesync-gpios = <0xffffffff 0x01 0x00>;
 			};
 		};
 	};
[...]

-		__overlay__ {
-
-			bt-timesync {
-				samsung,pins = "gpp11-1";
-				samsung,pin-function = <0x0f>;
-				samsung,pin-pud = <0x01>;
-				samsung,pin-drv = <0x00>;
-				samsung,pin-con-pdn = <0x03>;
-				samsung,pin-pud-pdn = <0x01>;
-				phandle = <0x08>;
-			};
-		};
-	};
-
[...]
 			__overlay__ {

 				btbcm {
-					pinctrl-0 = <0x00 0x04 0x08 0x0c>;
+					pinctrl-0 = <0x00 0x04 0x08>;
 				};
 			};
 		};

Bug: 360934165
Test: bluetooth still works; /proc/bluetooth/timesync still missing
Test: dtbo diff, per above

Change-Id: I685a8a95e4b80910f9fce24d898822e58bf76387
Signed-off-by: Brian Norris <briannorris@google.com>